3/3 ver2

+++ app/controllers/items_controller.rb
@@ -78,19 +78,21 @@ class ItemsController < ApplicationController
    def show
     @item = Item.find(params[:id])
     new_history = @item.histories.new
-    new_history.user_id = current_user.id
     
-    if current_user.histories.exists?(item_id: "#{params[:id]}")
-      old_history = current_user.histories.find_by(item_id: "#{params[:id]}")
-      old_history.destroy
-    end
+    if current_user.present?
+      new_history.user_id = current_user.id
+      if current_user.histories.exists?(item_id: "#{params[:id]}")
+        old_history = current_user.histories.find_by(item_id: "#{params[:id]}")
+        old_history.destroy
+      end
     
-    new_history.save
+      new_history.save
     
-    histories_stock_limit = 30
-    histories = current_user.histories.all
-    if histories.count > histories_stock_limit
-      histories[0].destroy
+      histories_stock_limit = 30
+      histories = current_user.histories.all
+      if histories.count > histories_stock_limit
+        histories[0].destroy
+      end
     end
    end
